+++ src/modules/graphic/opengl/src/debug/deoglDebugSnapshot.cpp
@@ -512,7 +512,7 @@ void deoglDebugSnapshot::pTakeSnapshotStates() const{
 }
 
 void deoglDebugSnapshot::pWriteLine( decBaseFileWriter &writer, const char *text ) const{
-	writer.Write( text, strlen( text ) );
+	writer.Write( text, ( int )strlen( text ) );
 	writer.WriteByte( '\n' );
 }
